Add time to brewing message
The status channel being mostly the same exact message makes my eyes
glaze over a bit when I check it, plus it takes a tiny amount of mental
effort to read the little timestamp and add five minutes to it. So this
changes the brewing message to say what time it will be ready.