Update double click message (#10)
* Update double click message

I think it's best to `@` the person who is responsible for purchasing new coffee when we are down to the last bag.

The double-click message seems much more useful to indicate that we are just out of brewed coffee. 

This PR just makes that change in wording for the double click.

* terser double click text

* Fix brittle test

Test is brittle because it looks for a specific string, which changed.